Added a corrected extinction divide by area of the first smps peak
The multiple charge approximation correction that we are doing to the
measured data is wrong. because it actually calculates sigma1*0.8=
sigmaT-sigma2*0.2. we are presenting sigma1*0.8, but what we actually
should be presenting is sigma1= (sigmaT-sigma2*0.2)/0.8. But, because
in our case we dealing with integrals we can't calculate this exactly,
unless we take the purely theoretical result, by normalizing the one
charge peak to one and then calculate the sigma1. We can approximate
instead and still used the measured result by calculating the constant
by taking the area of the first peak and then divide the correction
result by this number. then we can compare the delta function sigma
purely theoretical function to this approximated measured result.